+++ client/src/store/index.js
@@ -2,3 +2,5 @@ export { default as User } from './User';
 export { default as DateInfo } from './DateInfo';
 export { default as History } from './History';
 export { default as HistoryReport } from './HistoryReport';
+export { default as FormStore } from './Form';
+export { default as Payment } from './Payment';
